Figured out window trigger behavior! (Check commit message)
There is both a window X trigger and a window Y trigger. Window Y trigger is a prerequisite for the window X trigger. WY and LY are evaluated at the beginning of the scanline, and if they match, the window Y trigger is activated, enabling the window X trigger until Vblank.

This complex behavior is required to not break the battle intro in Pokemon Crystal. If the window X trigger is allowed to activate in any scanline, the visuals will glitch for 1 frame right as the slide-in animation ends.